ccms_backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysqlclient==2.2.8
sqlparse==0.5.5
